gmp.sh: clean after building
GMP has a very "noisy" build process in the sense that it creates a huge
number of output files that would need to be ignored by Git (not just .o
files but .c and .h files as well). It is simplest to just blow them
away rather than explicitly ignore them.

Signed-off-by: Wincent Colaiuta <win@wincent.com>