Don't check compiled Ruby into SCM
This is an intermediate build product, so it definitely shouldn't be
stored in SCM. Walrus itself now has a copy of the Hextrapolate
templates in its acceptance suite, so things are more than adequately
tested and there's not much point in keeping these lying around anymore.

In fact, even the generated HTML need not necessarily live in the SCM,
but will handle that in a later commit, if I decide to go ahead with
removing it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>