Avoid repeated `reverse()` calls in `getDigits` + `joinDigits`
Put most significant digits first, which means we avoid the `reverse()`
calls at the cost of having to use `unshift()`. It is indeed slower, but
I wanted to show my work. Next step will be replacing the `unshift()`
with `push()` + `reverse()`, but that probably still won't be enough to
get us back to perf parity.