fix: correct bits/bytes summary
We were taking numbers, converting them to hex, and counting their
length, and multiplying by 8:

    f -> 8 bits, 1 byte
    ff -> 16 bits, 2 bytes

One fix would be to multiply by 4 instead:

    f -> 4 bits, 0.5 bytes
    ff -> 8 bits, 1 byte

So this is correct, in a sense, but we don't really want to show
fractional bytes.

Another fix would be, then, to just add a `Math.ceil` so that we'd get:

    f -> 4 bits, 1 byte
    ff -> 8 bits, 1 byte

But that doesn't feel so good when you're typing binary numbers; eg.
even "1" would show up as "4 bits, 1 byte". By instead converting to
base-2 and continue to use the `Math.ceil` trick, we get the best of
both worlds for these binary values:

    1 -> 1 bit, 1 byte
    11 -> 2 bits, 1 byte
    111 -> 3 bits, 1 byte
    1111 -> 4 bits, 1 byte
    11111 -> 5 bits, 1 byte
    111111 -> 6 bits, 1 byte
    1111111 -> 7 bits, 1 byte
    11111111 -> 8 bits, 1 byte